% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcw_tests.R
\name{additional_testing}
\alias{additional_testing}
\title{Additional testing}
\usage{
additional_testing(hcws, screening_hcws, params, tests_weekly)
}
\arguments{
\item{tests_weekly}{}
}
\value{
Dataframe of additional testing
\describe{
  \item{week_ends}{xyz}
  \item{week_begins}{xyz}
  \item{tests_hcw_weekly}{xyz}
  \item{tests_contacts_weekly}{xyz}
}
}
\description{
Additional testing
}
