% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bed_nr_proxy}
\alias{bed_nr_proxy}
\title{Income group averages of hospital beds per 1,000}
\format{
A data frame with 5 rows and 3 variables:
\describe{
  \item{income_group}{Income group}
  \item{beds_1000}{Hospital beds per 1,000 people}
  \item{source}{Source of estimate}
}
}
\source{
\url{https://data.worldbank.org/indicator/SH.MED.BEDS.ZS}
}
\usage{
bed_nr_proxy
}
\description{
Average number of hospital beds per 1,000 people in a country population based on income group.
used as proxy numbers for when country level data not available.
}
\keyword{datasets}
