% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cases.R
\name{calculate_cases}
\alias{calculate_cases}
\title{Patient Calculations}
\usage{
calculate_cases(
  country_data,
  starting_date = "2022-01-02",
  forecast_length = 10,
  model = "Imperial SEIR",
  transmission_scenario = "Medium"
)
}
\arguments{
\item{transmission_scenario}{}
}
\value{

}
\description{
For now, this will work with imperial model outputs.
Add cumulative cases???
delivery_leadtime will likely have to go to forecasted equipment demand
}
