% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{capacity_perc}
\alias{capacity_perc}
\title{Percent capacity dedicated to COVID}
\format{
A data frame with 4 rows and 2 variables:
\describe{
  \item{capacity}{Diagnostic capacity dedicated to COVID-19 response}
  \item{percent}{Percentage of capacity this translates to}
}
}
\source{
\url{https://www.who.int/publications/i/item/WHO-2019-nCoV-Tools-Essential_forecasting-2022.1}
}
\usage{
capacity_perc
}
\description{
Percent capacity dedicated to COVID
}
\keyword{datasets}
