% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases_weekly.R
\name{cases_weekly}
\alias{cases_weekly}
\title{Produce Weekly Summary of Cases}
\usage{
cases_weekly(params, data, starting_date = "2019-11-29")
}
\arguments{
\item{starting_date}{}
}
\value{
Dataframe of weekly summary
\describe{
  \item{week_begins}{Date the week summarized begins}
  \item{week_ends}{Date the week summarized ends}
  \item{hospital_demand}{xyz}
  \item{ICU_demand}{xyz}
  \item{hospital_incidence}{xyz}
  \item{ICU_incidence}{xyz}
  \item{infections}{xyz}
  \item{cumulative_infections}{xyz}
  \item{cum_severe_cases}{xyz}
  \item{new_severe_cases}{xyz}
  \item{cum_critical_cases}{xyz}
  \item{new_critical_cases}{xyz}
  \item{adm_severe_cases_nocap}{xyz}
  \item{adm_critical_cases_nocap}{xyz}
  \item{adm_severe_cases_cap}{xyz}
  \item{adm_critical_cases_cap}{xyz}
  \item{new_mild_cases}{xyz}
  \item{new_mod_cases}{xyz}
  \item{new_mild_cases_2}{xyz}
  \item{new_mod_cases_2}{xyz}
  \item{new_severe_cases_2}{xyz}
  \item{new_critical_cases_2}{xyz}
  \item{cum_mild_cases}{xyz}
  \item{cum_mod_cases}{xyz}
  \item{rem_mild_cases}{xyz}
  \item{rem_mod_cases}{xyz}
  \item{rem_severe_cases}{xyz}
  \item{rem_critical_cases}{xyz}
  \item{cum_rem_mild_cases}{xyz}
  \item{cum_rem_mod_cases}{xyz}
  \item{cum_rem_severe_cases}{xyz}
  \item{cum_rem_critical_cases}{xyz}
}
}
\description{
Note: since proportion of mild and moderate are the same, the estimates of
these categories will be the same.
}
