% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diagnostics}
\alias{diagnostics}
\title{Diagnostics}
\format{
A data frame with 220 rows and 9 variables:
\describe{
  \item{country_name}{Country name}
  \item{who_region}{WHO region}
  \item{income_group}{Income group}
  \item{roche_6800}{Number of Roche 6800 machines estimated to be in country}
  \item{roche_8800}{Number of Roche 8800 machines estimated to be in country}
  \item{abbott_m2000}{Number of Abbott m2000 machines estimated to be in country}
  \item{hologic_panther}{Number of Hologic Panther machines estimated to be in country}
  \item{genexpert}{Number of GeneXpert machines estimated to be in country}
  \item{manual}{Number of manual machines estimated to be in country}
}
}
\source{
\url{https://apps.who.int/iris/bitstream/handle/10665/333983/WHO-2019-nCoV-Tools-Essential_forecasting-Overview-2020.1-eng.pdf?sequence=1&isAllowed=y}
}
\usage{
diagnostics
}
\description{
According to the WHo ESFT: "the reference values for the module were taken from an assessment of
available equipment and are based on a number of factors including population
size, HIV burden (as many machines were initially purchased for HIV testing),
and testing platforms known to the WHO. Please note that the reference values
are estimates and may not exactly match the number of platforms in each
country."
}
\details{
The high-throughput conventional machines are the Roche 6800, Roche 8800,
Abbott m2000, and the Hologic Panthers. These are also used for HIV testing.
GeneXpert machines are near-patient PCR processing machines, that provide on
the spot diagnostics for patients. Manual real time PCR platforms are slower
but also provide PCR processing.
}
\keyword{datasets}
