% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commodities_forecast.R
\name{diagnostics_forecast}
\alias{diagnostics_forecast}
\title{Diagnostics weekly}
\usage{
diagnostics_forecast(lab_params, equipment, test_ratios, n_tests, patients)
}
\arguments{
\item{lab_params}{From lab parameters}

\item{equipment}{This should be the data frame of equipment need}

\item{test_ratios}{This should be from test ratios}

\item{n_tests}{n_tests from total_tests}

\item{patients}{patients_weekly, this has num of hospital facilities in use}
}
\value{
Dataframe of diagnostics weekly
\describe{
  \item{category}{Category of item: diagnostics}
  \item{week_begins}{Date the week begins}
  \item{week_ends}{Date the week ends}
  \item{unit}{Unit items supplied in}
  \item{item}{Item name}
  \item{total_amount}{Total amount of that item}
}
}
\description{
Calculates a forecast of diagnostic needs by week.
}
