% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_weekly.R
\name{diagnostics_weekly}
\alias{diagnostics_weekly}
\title{Diagnostics Weekly}
\usage{
diagnostics_weekly(
  params,
  patients,
  cases,
  diagnostic_parameters,
  testing_strategy = "all"
)
}
\arguments{
\item{testing_strategy}{}
}
\value{
Dataframe of weekly summary
\describe{
  \item{week_begins}{xyz}
  \item{week_ends}{xyz}
  \item{tests_diagnosis_uncapped_sev_crit}{xyz}
  \item{tests_release_uncapped_sev_crit}{xyz}
  \item{tests_diagnosis_capped_sev_crit}{xyz}
  \item{tests_release_capped_sev_crit}{xyz}
  \item{tests_mild}{xyz}
  \item{tests_mod}{xyz}
  \item{tests_suspected}{xyz}
  \item{testing_strategy}{xyz}
}
}
\description{
This function takes some of the HCW cap options and calculates
the section in the `Weekly Summary` tab marked HCW and staff.
This will then be used in some sort of capacity mapping/forecasting.
}
