% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esft.R
\docType{package}
\name{esft}
\alias{esft}
\title{\code{esft} package}
\description{
WHO ESFT documentation
}
\details{
See the README on
Update when published:

\href{https://cran.r-project.org/package=esft/README.html}{CRAN}
or \href{https://github.com/mrc-ide/esft#readme}{GitHub}
}
