% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_capacity.R
\name{get_beds}
\alias{get_beds}
\title{Gets beds}
\usage{
get_beds(country_capacity, overrides = list())
}
\arguments{
\item{country_capacity}{a named list produced by get_country_capacity}

\item{overrides}{a named list of parameter values to use instead of defaults}
}
\value{
List of bed counts by category
}
\description{
Gets beds
}
