% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_caps.R
\name{get_caps}
\alias{get_caps}
\title{Calculate caps}
\usage{
get_caps(country = NULL, iso3c = NULL)
}
\arguments{
\item{iso3c}{}
}
\value{

}
\description{
Performs the back calculations to retrieve bed and hcw caps
}
