% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_capacity.R
\name{get_country_capacity}
\alias{get_country_capacity}
\title{Gets country capacity.}
\usage{
get_country_capacity(country = NULL, iso3c = NULL, overrides = list())
}
\arguments{
\item{country}{Country name}

\item{iso3c}{Country code, in iso3c format}

\item{overrides}{a named list of parameter values to use instead of defaults}
}
\value{
List of country name, iso3c code, and population.
}
\description{
Gets country capacity.
}
