% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_parameters.R
\name{get_country_parameters}
\alias{get_country_parameters}
\title{Sets country parameters.}
\usage{
get_country_parameters(country = NULL, iso3c = NULL, overrides = list())
}
\arguments{
\item{overrides}{}
}
\value{
List of country name, iso3c code, and population.
}
\description{
TO DO:
}
\details{
ALSO GET CUMULATIVE CASES???????
}
