% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_parameters.R
\name{get_country_test_capacity}
\alias{get_country_test_capacity}
\title{Get diagnostic capacity}
\source{
Estimates provided by the WHO Operations, Supply & Logistics Team and
reviewed by diagnostics technical experts at the WHO.
}
\usage{
get_country_test_capacity(iso3c = NULL, overrides = list())
}
\arguments{
\item{iso3c}{Country code, in iso3c format.}

\item{overrides}{a named list of parameter values to use instead of defaults.
Notably might include hologic_panther_fusion counts. The values are described
below and are taken from data provided in the ESFT.}
}
\value{
Number of diagnostic machines available within each country.
}
\description{
Using country name or country code, return baseline estimates of
diagnostic testing capacity provided in the WHO ESFT.
}
\details{
Counts of various machines in country:
\itemize{
 \item{roche_6800}{ - high throughput conventional platform}
 \item{roche_8800}{ - high throughput conventional platform}
 \item{abbott_m2000}{ - high throughput conventional platform}
 \item{hologic_panther}{ - high throughput conventional platform}
 \item{hologic_panther_fusion}{ - high throughput conventional platform}
 \item{genexpert}{ - near-patient PCR machine modules (not platforms: here,
 one machine contains several modules)}
 \item{manual}{ - manual real-time PCR platform}
}
}
