% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_parameters.R
\name{get_diagnostic_capacity}
\alias{get_diagnostic_capacity}
\title{Get diagnostic capacity}
\usage{
get_diagnostic_capacity(country = NULL, iso3c = NULL, overrides = list())
}
\arguments{
\item{country}{Country name.}

\item{iso3c}{Country code, in iso3c format.}

\item{overrides}{a named list of parameter values to use instead of defaults
The parameters are defined below.}
}
\description{
Using country name or country code, return baseline estimates of diagnostic
testing capacity provided in the WHO ESFT.
}
