% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_parameters.R
\name{get_diagnostic_parameters}
\alias{get_diagnostic_parameters}
\title{Get diagnostic parameters}
\usage{
get_diagnostic_parameters(overrides = list())
}
\arguments{
\item{overrides}{Named list of parameter values to use instead of defaults
The parameters are defined below.}
}
\description{
Sets the baseline diagnostic parameters.
}
