% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_capacity.R
\name{get_hcw_caps}
\alias{get_hcw_caps}
\title{Gets HCW caps}
\usage{
get_hcw_caps(country_capacity, params, overrides = list())
}
\arguments{
\item{country_capacity}{a named list produced by get_country_capacity}

\item{overrides}{a named list of parameter values to use instead of defaults}
}
\value{
List of HCW counts by category
}
\description{
Gets HCW caps
}
