% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_parameters.R
\name{get_lab_parmeters}
\alias{get_lab_parmeters}
\title{Get lab parameters}
\usage{
get_lab_parmeters(overrides = list())
}
\arguments{
\item{overrides}{Named list of parameter values to use instead of defaults
The parameters are defined below.}
}
\description{
Sets the baseline lab parameters.
}
