% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{get_parameters}
\alias{get_parameters}
\title{Get extra model parameters}
\usage{
get_parameters(overrides = list())
}
\arguments{
\item{overrides}{a named list of parameter values to use instead of defaults
The parameters are defined below.}
}
\value{
All of the misc parameters.
}
\description{

}
