% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcw_caps.R
\name{hcw_caps}
\alias{hcw_caps}
\title{HCW Caps}
\usage{
hcw_caps(params, capacity, throughput, hwfe, patients, overrides = list())
}
\arguments{
\item{params}{From get_parameters}

\item{capacity}{Country capacity, get_country_capacity}

\item{throughput}{Throughput dataframe, from r data file - can be altered}

\item{hwfe}{From WHO ESFT sheet}

\item{patients}{From patients_weekly}

\item{overrides}{a named list of hcw_cap values to use instead of defaults}
}
\value{
List of caps
\describe{
  \item{iso3c}{Iso3c code these HCW caps are associated with.}
  \item{bed_cap}{Hospital bed cap (equals number of beds per country, as
  found in the capacity function, which takes it from the World Bank)}
  \item{cases_screened_per_hcw_per_day}{Calculates the number of cases that
  could theoretically be screened per HCW per day based on standardized
  recommendations of hours spent per day per patient per HCW grouping (which
  was developed using HWFE tool methodology and consultation with clinical
  leads)}
  \item{cleaners_inpatient_cap}{Number of beds times the hygienists per bed
  in a hospital setting}
  \item{hcws_inpatients_cap}{Percent of HCWs treating COVID-19 patients times
  the number of HCWs}
  \item{hcws_per_bed}{Number of HCWs per bed, weighted by the total number of
  beds in use by severity and the time recommended for HCWs to spend at beds
  by severity}
  \item{hcws_per_inpatient}{Same as hcws_per_bed}
  \item{hcws_per_outpatient}{1/Number of cases screened per HCW per day}
  \item{hcws_screening_cap}{Percent of HCWs screening COVID-19 times the
  number of HCWs}
  \item{hygienists_per_bed}{Number of hygienists per bed, weighted by the
  total number of beds in use by severity and the time recommended for
  hygienists to spend at beds by severity}
  \item{lab_staff_cap}{If there are machines dedicated to COVID-19 test
  processing, its the average of their percent capacity allocation to COVID
  multiplied by the number of lab staff. If there are no estimates for
  percent allocation to COVID, it is equal to the total number of lab staff}
  \item{perc_crit_cases}{Percent of total critical beds in use divided by
  the total beds in use over the forecasting period}
  \item{perc_screening_covid}{Percent of HCWs dedicated to COVID-19
  screening, which is equal to the leftover Percent after accounting for
  those HCWs not dedicated to COVID-19 and those allocated to COVID-19
  response}
  \item{perc_sev_cases}{Percent of total severe beds in use divided by
  the total beds in use over the forecasting period}
  \item{perc_treating_covid}{Percent of HCWs dedicated to COVID-19
  response - this depends on the total patients in beds and the number of
  HCWs needed to respond to them during the forecasting period}
  \item{prob_inpatient}{Probability the patient in inpatient (severe or
  critical)}
  \item{prob_outpatient}{Probability the patient in inpatient (mild or
  moderate)}
  \item{ratio_hcws_inpatient_outpatient}{Ratio of inpatient dedicated HCWs
  to inpatient and outpatient HCWs}
}
}
\description{
This function calculates the HCW caps found in the
`Weekly Summary`, `Inputs`, and `Back Calculations` tabs in the ESFT.

Here is the description of the bed cap calculations:
\itemize{
 \item{bed_cap}{ - Bed cap to calculate max bed occupancy, equal to total
 number of beds available per country.}
 \item{cases_screened_per_hcw_per_day}{ - Screening/triage ratio is based on
 the assumption that each screening/triage takes approximately 48 minutes,
 which is 10 consultations per 8-hour shift.}
 \item{cleaners_inpatient_cap}{ - Max number of estimated cleaners needed,
 refers to Cleaners and Helpers (ILO ISCO code 9112).}
 \item{hcws_inpatients_cap}{ - Max number of estimated healthcare workers,
 refers to all medical practitioners, including physicians, nursing
 professionals, and paramedical practitioners (ILO ISCO codes 2240, 2211,
 2212, 2221, 3221, 5321, 3256)}
 \item{hcws_per_bed}{ - Average number of health care workers per bed,
 derived from bed occupancy and typical/recommended hours per bed per health
 care worker.}
 \item{hcws_per_inpatient}{ - Same as hcws_per_bed.}
 \item{hcws_per_outpatient}{ - Number of health care workers required per
 outpatient (screening) case, or the inverse of cases screened per health
 care worker per day.}
 \item{hcws_screening_cap}{ - Max number of estimated health care workers
 dedicated to screening COVID-19 cases.}
 \item{hygienists_per_bed}{ - Average number of hygienists per bed, derived
 from bed occupancy and typical/recommended hours per bed per hygienist.}
 \item{lab_staff_cap}{ - Estimated max number of possible lab staff - either
 derived from the different machines and their estimated COVID-19 capacity or
 derived from the total number of available lab staff.}
 \item{perc_crit_cases}{ - Percent of all patients occupying beds who are of
 critical severity.}
 \item{perc_screening_covid}{ - Percent of all HCWs dedicated to screening
 COVID-19 cases (usually the percent leftover after accounting for those who
 are and are not dedicated to COVID-19 response)}
 \item{perc_sev_cases}{ - Percent of all patients occupying beds who are of
 severe case severity.}
 \item{perc_treating_covid}{ - Percent of HCWs dedicated to responding to
 COVID-19.}
 \item{prob_inpatient}{ - Probability that a COVID-19 case is inpatient, i.e.
 in a care facility.}
 \item{prob_outpatient}{ - Probability that a COVID-19 case is outpatient,
 i.e. that it is not in a care facility.}
 \item{ratio_hcws_inpatient_outpatient}{ - Per new case, ratio of HCWs for
 inpatient vs. outpatient}
}
}
