% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcw_caps.R
\name{hcw_caps_dynamic}
\alias{hcw_caps_dynamic}
\title{Dynamic HCW caps}
\usage{
hcw_caps_dynamic(
  params,
  hwfe,
  patients,
  ambulanciers_per_bed = 0.06,
  bio_eng_per_bed = 0.02
)
}
\arguments{
\item{params}{From get_country_capacity}

\item{hwfe}{From WHO ESFT sheet}

\item{patients}{From patients_weekly}

\item{ambulanciers_per_bed}{Assumes 1 ambulance per 100 bed hospital w 2
operators at all times (3x8 hour shifts) = (2/100)*3 = 0.06, taken from ESFT}

\item{bio_eng_per_bed}{Assumes 2 biomed engineers on 8 hr shifts per 100 bed
hospital, taken from ESFT}
}
\value{
Dataframe of weekly summary
\describe{
  \item{hygienists_per_bed}{Date the week summarized begins}
  \item{hygienists_per_crit_bed}{Date the week summarized ends}
  \item{cleaners_inpatient_cap}{Number cleaners required based on hospital
  beds in use}
  \item{inf_caregiver_inpatient_cap}{Number caregivers required based on
  hospital beds in use}
  \item{amb_personnel_inpatient_cap}{Ambulanciers required based on hospital
  beds in use}
  \item{bio_eng_inpatient_cap}{Biomedical engineers required based on
  hospital beds in use}
}
}
\description{
Dynamic HCW caps
}
