% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcw_caps.R
\name{hcw_caps_static}
\alias{hcw_caps_static}
\title{Static HCW Caps}
\usage{
hcw_caps_static(params, capacity, throughput, cap_lab_staff = FALSE)
}
\arguments{
\item{params}{From get_parameters}

\item{capacity}{Country capacity, get_country_capacity}

\item{throughput}{Throughput dataframe, from r data file - can be altered}

\item{cap_lab_staff}{TRUE/FALSE. Option to set a cap on lab staff based on
percent of diagnostic machinery allocated to COVID}
}
\value{
List of caps
\describe{
  \item{hcws_inpatients_cap}{Date the week summarized begins}
  \item{hcws_screening_cap}{Date the week summarized ends}
  \item{lab_staff}{Lab staff available}
  \item{cap_lab_staff}{TRUE/FALSE, whether you want to cap lab staff by
  machines allocated to COVID or not}
}
}
\description{
This function calculates the static HCW caps found in the
`Weekly Summary` in the ESFT.
}
