% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_parameters.R
\docType{data}
\name{high_transmission}
\alias{high_transmission}
\title{Preset parameters for the high transmission scenario}
\format{
An object of class \code{character} of length 4.
}
\usage{
high_transmission
}
\description{
Preset parameters for the high transmission scenario
}
\details{
Use a vector of preset parameters for the high transmission scenario

Default parameters are Death Calibrated Label = Increase
Transmission 50%,
Not Death Calibrated Label = High Transmission, R(eff) or R(0) = 1.41,
Imperial Category Label = Relax Interventions 50%
}
\keyword{datasets}
