% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hours_per_shift}
\alias{hours_per_shift}
\title{Hours per shift number for diagnostic machines}
\format{
A data frame with 3 rows and 2 variables:
\describe{
  \item{shifts}{Number of shifts per day}
  \item{hours}{Total hours working with this number of shifts}
}
}
\source{
\url{https://www.who.int/publications/i/item/WHO-2019-nCoV-Tools-Essential_forecasting-2022.1}
}
\usage{
hours_per_shift
}
\description{
Hours per shift number for diagnostic machines
}
\keyword{datasets}
