% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hwfe}
\alias{hwfe}
\title{Health care work force data}
\format{
A data frame with 21 rows and 7 variables:
\describe{
  \item{occupational_title}{Occupational title}
  \item{patient_t24_mild}{Patient time per 24 hrs per mild case}
  \item{patient_t24_mod}{Patient time per 24 hrs per moderate case}
  \item{patient_t24_sev}{Patient time per 24 hrs per severe case}
  \item{patient_t24_crit}{Patient time per 24 hrs per critical case}
  \item{patient_t24_screen}{Patient time per 24 hrs for screening/triage}
  \item{esft_group}{Category of worker in ESFT}
}
}
\source{
\url{https://www.researchgate.net/figure/Health-Workforce-Estimator-HWFE-tool-applies-the-WISN-approach-to-caring-for-COVID_fig3_358174845}
}
\usage{
hwfe
}
\description{
HWFE tool methodology, combined with LMIC-specific inputs from
consultations, e.g., with Ethiopian clinical leads. Shows patient time per
24 hours by severity type.
}
\details{
Also uses adaptt tool:
\url{https://www.euro.who.int/en/health-topics/Health-systems/pages/strengthening-the-health-system-response-to-covid-19/surge-planning-tools/adaptt-surge-planning-support-tool}
These tools yield different outputs based on different focus + models. Note
ESFT has SEIR models, other tools use simplified projections/epi models.
These are simplified groups from HWC Need worksheet (external). Patient time
is given across different settings, according to staff skills and scope of
practice. Note that only hours/day is used in the calculations.
}
\keyword{datasets}
