% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icl_data}
\alias{icl_data}
\title{Imperial Model Fits to Excess Death Data}
\format{
A data frame with 3,298,878 rows and 6 variables:
\describe{
  \item{scenario}{Transmission scenario, as given in the ESFT tool.}
  \item{compartment}{Projection compartment.}
  \item{date}{Date.}
  \item{death_calibrated}{Logical: Calibrated to reported deaths (TRUE) or
  to excess deaths (FALSE).}
  \item{y_mean}{Mean value of compartment specified.}
  \item{iso3c}{Iso3c country codes.}
}
}
\source{
\url{https://github.com/mrc-ide/global_lmic_projections_esft}
}
\usage{
icl_data
}
\description{
Imperial Model Fits to Excess Death Data
}
\keyword{datasets}
