% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_parameters.R
\docType{data}
\name{low_transmission}
\alias{low_transmission}
\title{Preset parameters for the low transmission scenario}
\format{
An object of class \code{character} of length 4.
}
\usage{
low_transmission
}
\description{
Preset parameters for the low transmission scenario
}
\details{
Use a vector of preset parameters for the low transmission scenario

Default parameters are Death Calibrated Label = Decrease Transmission 50%,
Not Death Calibrated Label = Low Transmission, R(eff) or R(0) = 0.47, Imperial Category Label = Additional 50% Reduction
}
\keyword{datasets}
