% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_weekly.R
\name{max_tests_per_day}
\alias{max_tests_per_day}
\title{Max tests per day}
\usage{
max_tests_per_day(diagnostic_capacity)
}
\arguments{
\item{diagnostic_capacity}{from calc_diagnostic_capacity}
}
\description{
Uses the country specific diagnostic capacity estimates to
calculate the max number of tests per day.
}
