% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_parameters.R
\docType{data}
\name{med_transmission}
\alias{med_transmission}
\title{Preset parameters for the medium transmission scenario}
\format{
An object of class \code{character} of length 4.
}
\usage{
med_transmission
}
\description{
Preset parameters for the medium transmission scenario
}
\details{
Use a vector of preset parameters for the medium transmission scenario

Default parameters are Death Calibrated Label = Maintain Current Transmission,
Not Death Calibrated Label = Medium Transmission, R(eff) or R(0) = 0.94, Imperial Category Label = Maintain Status Quo
}
\keyword{datasets}
