% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{noncovid}
\alias{noncovid}
\title{Non-COVID Essentials}
\format{
A data frame with 6 rows and 23 variables.
\describe{
  \item{category}{Equipment type category: testing, infection prevention &
  control (IPC), case management - biomedical equipment, or case management
  - accessories & consumables.}
  \item{group}{Sub-categories of groups within each category.}
  \item{item}{Item name, with details}
  \item{unit}{Unit of measurement of the item}
  \item{reusable}{TRUE/FALSE: Is this item reusable or not.}
  \item{amount_per_noncovid_doctor_per_day}{Amount required per medical
  doctor not associated with the COVID-19 response per day.}
  \item{amount_per_noncovid_nurse_per_day}{Amount required per nurse
  not associated with the COVID-19 response per day.}
  \item{amount_per_noncovid_lab_tech_per_day}{Amount required per laboratory
  scientist and/or technician not associated with the COVID-19 response per
  day.}
  \item{amount_per_noncovid_midwife_per_day}{Amount required per midwife per
  day.}
  \item{amount_per_noncovid_dentist_per_day}{Amount required per dentist per
  day.}
  \item{amount_per_noncovid_physio_per_day}{Amount required per
  physiotherapist per day.}
  \item{amount_per_noncovid_traditional_compl_per_day}{Amount required per
  traditional and complementary medical personnel per day.}
  \item{amount_per_noncovid_chw_per_day}{Amount required per community health
  worker (CHW) per day.}
  \item{amount_per_noncovid_pharmacist_per_day}{Amount required per
  pharmacist per day.}
}
}
\usage{
noncovid
}
\description{
Non-COVID Essentials
}
\keyword{datasets}
