% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharma_forecast.R
\name{pharma_forecast}
\alias{pharma_forecast}
\title{Pharmaceuticals Forecast}
\usage{
pharma_forecast(pharmaceuticals, cases)
}
\description{
Takes the pharma inputs and times them by total new cases.
Unclear how to work out the price setting aspect.
}
