% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population}
\alias{population}
\title{UNDP Population data}
\format{
A data frame with 259 rows and 6 variables:
\describe{
  \item{country_wb}{Country name in World Bank data.}
  \item{country_undp}{Country name in UNDP data.}
  \item{country_code}{Country code, iso3c format.}
  \item{pop_2020}{Population of country in 2020.}
  \item{pop_2015}{Population of country in 2015.}
  \item{yoy}{Year over year growth of country population.}
}
}
\source{
\url{https://population.un.org/wup/Download/}
}
\usage{
population
}
\description{
UNDP Population data
}
\keyword{datasets}
