% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_imperial_data.R
\name{read_url}
\alias{read_url}
\title{Reads RDS files from urls}
\source{
\url{https://stackoverflow.com/questions/12193779/how-to-write-trycatch-in-r}
}
\usage{
read_url(url)
}
\arguments{
\item{url}{The url of the Rds file of the country fit from github.}
}
\value{
Whatever url is read.
}
\description{
Reads RDS files from urls
}
