% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_imperial_data.R
\name{readUrl}
\alias{readUrl}
\title{Reads RDS files from urls}
\usage{
readUrl(url)
}
\arguments{
\item{url}{The url of the Rds file of the country fit from github.}
}
\value{
Whatever url is read.
}
\description{
Reads RDS files from urls
}
