% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcw_tests.R
\name{screening_hcw_weekly}
\alias{screening_hcw_weekly}
\title{Screening HCWs per week, capped and uncapped}
\usage{
screening_hcw_weekly(diagnostics_weekly, params)
}
\arguments{
\item{params}{}
}
\value{
Dataframe of summary
\describe{
  \item{week_ends}{xyz}
  \item{week_begins}{xyz}
  \item{screening_hcw_uncapped}{xyz}
  \item{screening_hcw_capped}{xyz}
}
}
\description{
Screening HCWs per week, capped and uncapped
}
