% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcw_tests.R
\name{screening_hcws_weekly}
\alias{screening_hcws_weekly}
\title{Screening HCWs per week, capped and uncapped}
\usage{
screening_hcws_weekly(diagnostics_weekly, hcw_caps, capacity)
}
\arguments{
\item{diagnostics_weekly}{From diagnostics_weekly}

\item{hcw_caps}{From hcw_caps}

\item{capacity}{Country capacity, get_country_capacity}
}
\value{
Dataframe of summary
\describe{
  \item{week_begins}{Date the week summarized begins}
  \item{week_ends}{Date the week summarized ends}
  \item{screening_hcw_uncapped}{Number of HCWs required for screening/triage,
  based on tests needed for outpatients (i.e. suspected negatives, mild and
  moderate cases) per day, and number of cases one HCW can screen a day}
  \item{screening_hcw_capped}{Number of HCWs allocated to screening/triage:
  minimum of uncapped HCWs as calculated above and number of HCWs available/
  reported per country and the percentage of HCWs allocated to screening,
  which is a parameter that can be modified in get_parameters}
}
}
\description{
Screening HCWs per week, capped and uncapped
}
