% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_parameters.R
\name{set_scenarios}
\alias{set_scenarios}
\title{Parameterise scenarios to use in the model}
\usage{
set_scenarios(parameters, scenarios)
}
\arguments{
\item{parameters}{the model parameters}

\item{scenarios}{a list of scenario parameters, can be set using presets}
}
\description{
Parameterise scenarios to use in the model
}
\details{
Creates a dataframe of scenarios with their associated parameters
}
