% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_parameters.R
\name{set_testing_strategy}
\alias{set_testing_strategy}
\title{Sets testing strategy and associated parameters}
\usage{
set_testing_strategy(
  strategy = NULL,
  perc_tested_mild_mod = NULL,
  overrides = list()
)
}
\arguments{
\item{overrides}{a named list of parameter values to use instead of defaults
The parameters are defined below.}
}
\description{
Default is testing strategy = all
}
