% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_parameters.R
\name{total_labs}
\alias{total_labs}
\title{Calculates max total labs that could be available for COVID}
\usage{
total_labs(capacity)
}
\arguments{
\item{capacity}{From get_diagnostic_capacity. Or
calculate_diagnostic_capacity. Only thing is, we need the country capacity.}
}
\description{
Calculates max total labs that could be available for COVID
}
