% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcw_tests.R
\name{total_tests}
\alias{total_tests}
\title{Total tests}
\usage{
total_tests(tests_weekly, additional_tests, max_tests)
}
\arguments{
\item{max_tests}{}
}
\value{
Total tests
\describe{
  \item{week_ends}{xyz}
  \item{week_begins}{xyz}
  \item{tests_tests_capped}{xyz}
}
}
\description{
Total tests
}
