% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transmission_scenarios}
\alias{transmission_scenarios}
\title{Transmission scenarios}
\format{
A data frame with 3 rows and 5 variables:
\describe{
  \item{imperial_scenario}{Scenario as given in the ESFT tool.}
  \item{label_death_calibrated}{Label of the scenario in the death calibrated fits.}
  \item{label_not_death_calibrated}{Label of the scenario in the non death calibrated fits.}
  \item{R}{R number associated with transmission scenarios.}
  \item{imperial_category_labels}{Scenario labels as used in imperial model fits.}
}
}
\source{
\url{https://www.who.int/publications/i/item/WHO-2019-nCoV-Tools-Essential_forecasting-2022.1}
}
\usage{
transmission_scenarios
}
\description{
Transmission scenarios
}
\keyword{datasets}
