% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_input.R
\name{user_input}
\alias{user_input}
\title{User input}
\usage{
user_input(overrides = list())
}
\arguments{
\item{overrides}{a named list of parameter values to use instead of defaults
The parameters are defined below.}
}
\description{
Given that these are just three - don't know if these are useful.
Basically this is what is in the input/user dashboard tab that isn't captured
through other parameter-setting functions.
}
