% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wb_beds}
\alias{wb_beds}
\title{World Bank estimates of number of hospital beds per country}
\format{
A data frame with 264 rows and 65 variables:
\describe{
  \item{country_name}{Country name}
  \item{country_code}{Iso3c codes}
  \item{indicator_name}{Indicator name, in this case, hospital beds per 1,000
  people.}
  \item{indicator_code}{Indicator code,in this case, SH.MED.BEDS.ZS.}
  \item{1960...2019}{Years}
  \item{most_recent}{Most recent bed value: either reported, or income group
  average if not available.}
}
}
\source{
\url{https://data.worldbank.org/indicator/SH.MED.BEDS.ZS}
}
\usage{
wb_beds
}
\description{
Number of hospital beds per 1,000 people by country by year. The
most recent year of data varies across countries.
}
\keyword{datasets}
