% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{who_data}
\alias{who_data}
\title{WHO Coronavirus (COVID-19) Dashboard}
\format{
A data frame with 305,019 rows and 8 variables:
\describe{
  \item{Date_reported}{Date of reporting to WHO}
  \item{Country_code}{ISO Alpha-2 country code}
  \item{Country}{Country, territory, area}
  \item{WHO_region}{	WHO regional offices: WHO Member States are grouped into
  six WHO regions -- Regional Office for Africa (AFRO), Regional Office for
  the Americas (AMRO), Regional Office for South-East Asia (SEARO), Regional
  Office for Europe (EURO), Regional Office for the Eastern Mediterranean
  (EMRO), and Regional Office for the Western Pacific (WPRO).}
  \item{New_cases}{New confirmed cases. Calculated by subtracting previous
  cumulative case count from current cumulative cases count.}
  \item{Cumulative_cases}{	Cumulative confirmed cases reported to WHO to
  date.}
  \item{New_deaths}{New confirmed deaths. Calculated by subtracting previous
  cumulative deaths from current cumulative deaths}
  \item{Cumulative_deaths}{Cumulative confirmed deaths reported to WHO to
  date.}
}
}
\source{
\url{https://covid19.who.int/data}
}
\usage{
who_data
}
\description{
WHO Coronavirus (COVID-19) Dashboard
}
\keyword{datasets}
