% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_parameters.R
\name{calc_diagnostic_capacity}
\alias{calc_diagnostic_capacity}
\title{Given specified parameters, calculates diagnostic country capacity.}
\usage{
calc_diagnostic_capacity(
  country_diagnostic_capacity,
  throughput,
  hours_per_shift
)
}
\arguments{
\item{country_diagnostic_capacity}{Capacity called from the
get_country_test_capacity function}

\item{throughput}{Throughput data, loaded in package, from ESFT}

\item{hours_per_shift}{Hours per shift, default = 8}
}
\description{
Given specified parameters, calculates diagnostic country capacity.
}
