% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{capacity_perc}
\alias{capacity_perc}
\title{Capacity as translated to percent}
\format{
A data frame with 4 rows and 2 variables:
\describe{
  \item{capacity}{Capacity is string form}
  \item{percent}{What the string translated to as percent}
}
}
\usage{
capacity_perc
}
\description{
Capacity as translated to percent
}
\keyword{datasets}
