% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commodities_forecast.R
\name{case_management_forecast}
\alias{case_management_forecast}
\title{Case management weekly: accessories, consumables, and biomedical equipment}
\usage{
case_management_forecast(params, equipment, patients)
}
\arguments{
\item{params}{Classic parameter function (?)}

\item{equipment}{This should be the data frame of equipment need}

\item{patients}{From patients_weekly}
}
\description{
in O(N^2) time. Next iteration should get rid of for loops.
To do: get rid of for loop for non reusable items.
Can also experiment with Reduce.
}
