% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases_weekly.R
\name{cases_weekly}
\alias{cases_weekly}
\title{Produce Weekly Summary of Cases}
\usage{
cases_weekly(
  params,
  capacity,
  test_strategy_params,
  data,
  data_source = "Imperial",
  user
)
}
\arguments{
\item{params}{Includes both estimates of beds and case severity proportions.}

\item{capacity}{From get_country_capacity}

\item{test_strategy_params}{From set_testing_strategy}

\item{data}{Specific country fit data, from the imperial model fits or from
WHO data}

\item{data_source}{Either WHO or Imperial.}

\item{user}{From user_input function}
}
\value{
Dataframe of weekly summary, if data_source = "Imperial". If the
data_source = "WHO", the ICU and hospital columns will not be included.
\describe{
  \item{week_begins}{Date the week summarized begins}
  \item{week_ends}{Date the week summarized ends}
  \item{hospital_demand}{Summed hospital demand: number of people who would
  be using a hospital bed given enough healthcare capacity but won't
  necessarily have it}
  \item{ICU_demand}{Summed ICU demand: number of people who would be using a
  hospital bed given enough healthcare capacity but won't necessarily have
  it}
  \item{hospital_incidence}{Summed hospital incidence: the total number of
  new people who need a new hospital bed at the current time. This does not
  include those recovering from ICU in a hospital bed or who already have
  access to a bed.}
  \item{ICU_incidence}{Summed ICU incidence: the total number of
  new people who need a new ICU bed at the current time. This does not
  include those who already have access to a bed.}
  \item{infections}{Estimated number of new infections, from model fits}
  \item{cumulative_infections}{Cumulative number of infections}
  \item{new_critical_cases}{For Imperial input = new critical cases per week
  as defined by as the ICU incidence. For WHO input, this is equal to the
  number of new cases per week multiplied by the proportion of infections
  that are critical}
  \item{new_severe_cases}{For Imperial input = new severe cases per week
  as defined by as the hospital incidence. For WHO input, this is equal to
  the number of new cases per week multiplied by the proportion of infections
  that are severe}
  \item{new_mod_cases}{New moderate cases every week, found by doing a
  transformation of the critical and severe cases and multiplying by the
  moderate case proportion}
  \item{new_mild_cases}{New mild cases every week, found by doing a
  transformation of the critical and severe cases and multiplying by the mild
  case proportion}
  \item{cum_critical_cases}{Cumulative critical cases per week as defined by
  the ESFT: cumulative ICU incidence}
  \item{cum_severe_cases}{Cumulative severe cases per week as defined by the
  ESFT: cumulative hospital incidence}
  \item{cum_mod_cases}{Cumulative moderate cases, using the first moderate
  case calculation}
  \item{cum_mild_cases}{Cumulative mild cases, using the first mild case
  calculation}
  \item{sus_cases_but_negative}{Sum of all new cases multiplied by the
  number of negative tests per positive case}
}
}
\description{
Replicates the calculations in the 'Weekly Summary' tab of the
ESFT depending on the input data source (WHO or Imperial). There are elements
of these calculations, such as the assumption that hospital incidence =
number of severe cases in the Imperial calculations, that the ICL team does
not make in its calculations. Therefore, keep in mind that this is just a
duplicate of the calculations in the ESFT, and not a critical appraisal of
these calculations. This calculation does not yield an output where the first
row contains the start date, as rows previous to the start date are needed
for the patient calculations.

Note: since proportion of mild and moderate are the same, the estimates of
these categories will be the same.
}
