% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commodities_forecast.R
\name{diagnostics_forecast}
\alias{diagnostics_forecast}
\title{Diagnostics weekly}
\usage{
diagnostics_forecast(lab_params, equipment, test_ratios, total_tests, patients)
}
\arguments{
\item{lab_params}{From lab parameters}

\item{equipment}{This should be the data frame of equipment need}

\item{test_ratios}{This should be from test ratios}

\item{total_tests}{total_tests from that function}

\item{patients}{patients_weekly, this has the number of hospital facilities in use}
}
\description{

}
