% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_weekly.R
\name{diagnostics_weekly}
\alias{diagnostics_weekly}
\title{Diagnostics Weekly}
\usage{
diagnostics_weekly(
  params,
  patients,
  cases,
  diagnostic_parameters,
  testing_scenario
)
}
\arguments{
\item{params}{From get_parameters}

\item{patients}{From patients_weekly}

\item{cases}{From cases_weekly}

\item{diagnostic_parameters}{From get_diagnostic_parameters}

\item{testing_scenario}{List of 8 testing scenario parameters, from
set_testing_strategy()}
}
\value{
Dataframe of weekly summary
\describe{
  \item{week_begins}{Date the week summarized begins}
  \item{week_ends}{Date the week summarized ends}
  \item{tests_diagnosis_uncapped_sev_crit}{Tests required if all severe and
  critical patients tested for diagnosis, regardless of capacity limits.
  Number of tests required per case specified in diagnostic_parameters.}
  \item{tests_release_uncapped_sev_crit}{Tests performed if all severe and
  critical patients tested for release, regardless of capacity limits.
  Number of tests per case specified in diagnostic_parameters.}
  \item{tests_diagnosis_capped_sev_crit}{Tests for diagnosis for severe
  and critical patients capped by hospital capacity/number of severe and
  critical patients admitted to hospital}
  \item{tests_release_capped_sev_crit}{Tests for release for severe and
  critical patients capped by hospital capacity and IFR/number of severe and
  critical patients discharged from hospital}
  \item{tests_mod}{Number of tests for moderate patient diagnosis, given
  testing strategy}
  \item{tests_mild}{Number of tests for mild patient diagnosis, given testing
  strategy}
  \item{tests_suspected}{Number of tests for suspected but negative cases}
  \item{testing_strategy}{Testing strategy, string}
}
}
\description{
This function takes some of the HCW cap options and calculates
the section in the `Weekly Summary` tab marked HCW and staff.
This will then be used in some sort of capacity mapping/forecasting.
}
