% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_capacity.R
\name{get_country_capacity}
\alias{get_country_capacity}
\title{Gets country capacity.}
\source{
\url{https://population.un.org/wup/Download/}

\url{https://www.who.int/publications/i/item/WHO-2019-nCoV-Tools-Essential_forecasting-2022.1}

\url{https://data.worldbank.org/indicator/SH.MED.BEDS.ZS}

Imperial College, Report 12: The Global Impact of COVID-19 and Strategies for Mitigation and Suppression

\url{https://www.researchgate.net/figure/Health-Workforce-Estimator-HWFE-tool-applies-the-WISN-approach-to-caring-for-COVID_fig3_358174845}
}
\usage{
get_country_capacity(country = NULL, iso3c = NULL, overrides = list())
}
\arguments{
\item{country}{Country name (in conventional capitalized format)}

\item{iso3c}{Country code, in iso3c format (in all capital letters)}

\item{overrides}{a named list of parameter values to use instead of defaults}
}
\value{
List of country capacity parameters.
\describe{
  \item{country}{Country name}
  \item{iso3c}{Country code, iso3c format}
  \item{population}{xyz}
  \item{yoy_growth}{Year over year growth of country population.}
  \item{income_group}{Income group}
  \item{n_hcws}{Number of HCWs (sum nurses + doctors, from HWFE data)}
  \item{n_labs}{Number of lab staff (from HWFE data)}
  \item{n_hosp_beds}{Est beds per country, from World Bank}
  \item{perc_beds_crit_covid}{Percent beds for ICU patients per country,
  estimated from World Bank/Imperial College Report}
  \item{perc_beds_not_covid}{Percent beds not allocated to COVID, default
  = 0.4}
  \item{perc_beds_sev_covid}{Percent beds allocated to severe COVID- cases,
  1 - (perc_beds_crit_covid + perc_beds_not_covid)}
  \item{beds_covid}{Number of hospital beds allocated to COVID (num hospital
  beds * perc allocated to covid)}
  \item{severe_beds_covid}{Number of severe beds allocated to COVID}
  \item{crit_beds_covid}{Number of critical beds allocated to COVID}
}
}
\description{
Note: put either country name or country code, and follow the
countrycode package convention on capitalization.
}
