% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_parameters.R
\name{get_diagnostic_parameters}
\alias{get_diagnostic_parameters}
\title{Get diagnostic parameters}
\usage{
get_diagnostic_parameters(overrides = list())
}
\arguments{
\item{overrides}{Named list of values to use instead of defaults
The parameters are defined below, and are taken from the default settings in
the ESFT.

total tests and percent antigen:

* total_tests_mild_mod - total tests, used for diagnosis only; default = 1
* total_tests_sev_crit - total tests, used for diagnosis and release;
default = 2
* perc_antigen_tests - percent testing done in hospital via antigen testing,
max 80%; default = 20%}
}
\value{
List of diagnostic parameters.
}
\description{
Sets the baseline diagnostic parameters.
}
