% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{get_parameters}
\alias{get_parameters}
\title{Get model parameters}
\usage{
get_parameters(overrides = list())
}
\arguments{
\item{overrides}{a named list of parameter values to use instead of defaults
The parameters are defined below.

fixed state transitions:

* dd - the delay for humans to move from state D to A; default = 5
* dt - the delay for humans to move from state Tr to Ph; default = 5
* da - the delay for humans to move from state A to U; default = 200

#' treatment parameters:
please set treatment parameters with the convenience functions in
`drug_parameters.R`}
}
\description{
get_parameters creates a named list of parameters for use in the model. These
parameters are passed to process functions. These parameters are explained in
"The US President's Malaria Initiative, Plasmodium falciparum transmission
and mortality: A modelling study."
}
