% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcws_weekly.R
\name{hcws_weekly}
\alias{hcws_weekly}
\title{HCWS Weekly}
\usage{
hcws_weekly(
  params,
  lab_params,
  tests,
  patients,
  t_labs,
  hcw_dyn_caps,
  hcw_stat_caps
)
}
\arguments{
\item{lab_params}{get_lab_parameters}

\item{tests}{diagnostics_weekly}

\item{patients}{patients_weekly}

\item{t_labs}{From total_labs}

\item{hcw_dyn_caps}{dataframe of dynamic caps, hcw_caps_dynamic}

\item{hcw_stat_caps}{single list of caps, hcw_caps_static}
}
\value{
Dataframe of summary
\describe{
  \item{week_ends}{xyz}
  \item{week_begins}{xyz}
  \item{hcws_inpatient_capped}{xyz}
  \item{hcws_inpatient_uncapped}{xyz}
  \item{cleaners_inpatient_capped}{xyz}
  \item{amb_personnel_inpatient_capped}{xyz}
  \item{bio_eng_inpatient_capped}{xyz}
  \item{inf_caregivers_isol_uncapped}{xyz}
  \item{lab_staff_capped}{xyz}
  \item{cleaners_lab}{xyz}
}
}
\description{
This function takes some of the HCW cap options and calculates
the section in the `Weekly Summary` tab marked HCW and staff.
This will then be used in some sort of capacity mapping/forecasting.
}
