% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noncovid_essentials.R
\name{noncovid_essentials}
\alias{noncovid_essentials}
\title{Non COVID essentials}
\usage{
noncovid_essentials(noncovid, ref_hcws, forecast_length = 12, days_week = 5)
}
\arguments{
\item{noncovid}{Data frame of equipment need, noncovid.rda}

\item{ref_hcws}{Returned by reference_hcw}

\item{forecast_length}{Length of forecast, in number of weeks (default=12)}

\item{days_week}{Average days worked per week, default = 5}
}
\description{
Non COVID essentials
}
