% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pharmaceuticals}
\alias{pharmaceuticals}
\title{Pharmaceuticals}
\format{
A data frame with 147 rows and 23 variables:
\describe{
  \item{covid_specific}{Whether or not the drug is specific to COVID-19
  treatment - TRUE if COVID-19 specific, FALSE if not.}
  \item{drug}{Drug product}
  \item{classification}{Drug family to which the drug product belongs}
  \item{concentration}{Dosage and drug form}
  \item{formulation}{Dosage only}
  \item{units}{Unit in which the dosage is measured}
  \item{drug_form}{Formulation in which the drug is packages (i.e. tablet)}
  \item{price_usd}{Price per unit of drug product}
  \item{num_days_per_crit_treatment}{Number of days per treatment course for
  critical patients}
  \item{daily_amount_crit}{Daily amount for critical patients}
  \item{form_per_crit_course}{Total drug form per treatment course per
  critical patients}
  \item{perc_crit_treated}{Percentage of critical patients receiving this
  treatment}
  \item{num_days_per_sev_treatment}{Number of days per treatment course for
  severe patients}
  \item{daily_amount_sev}{Daily amount for severe patients}
  \item{form_per_sev_course}{Total drug form per treatment course per severe
  patients}
  \item{perc_sev_treated}{Percentage of severe patients receiving this
  treatment}
  \item{num_days_per_mod_treatment}{Number of days per treatment course for
  moderate patients}
  \item{daily_amount_mod}{Daily amount for moderate patients}
  \item{form_per_mod_course}{Total drug form per treatment course per
  moderate patients}
  \item{perc_mod_treated}{Percentage of moderate patients receiving this
  treatment}
  \item{num_days_per_mild_treatment}{Number of days per treatment course for
  mild patients}
  \item{daily_amount_mild}{Daily amount for mild patients}
  \item{form_per_mild_course}{Total drug form per treatment course per mild
  patients}
  \item{perc_mild_treated}{Percentage of mild patients receiving this
  treatment}
}
}
\source{
\url{https://www.who.int/publications/i/item/WHO-2019-nCoV-Tools-Essential_forecasting-2022.1}
}
\usage{
pharmaceuticals
}
\description{
Data frame of all pharmaceutical commodities that are included
in the forecast. The items included are based on bundles recommended by the
WHO and estimates are generated from scenario patient volumes by patient
type.
}
\keyword{datasets}
