% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commodities_forecast.R
\name{ppe_forecast}
\alias{ppe_forecast}
\title{PPE need weekly}
\usage{
ppe_forecast(equipment, hcws, patients, cases, tests, screening_hcws)
}
\arguments{
\item{equipment}{This should be the data frame of equipment need}

\item{hcws}{HCWs_weekly}

\item{patients}{patients_weekly}

\item{cases}{cases_weekly}

\item{tests}{diagnostics_weekly}

\item{screening_hcws}{screening_hcws_weekly}
}
\description{
PPE need weekly
}
