% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_imperial_data+greg.R,
%   R/load_imperial_data.R
\name{readUrl}
\alias{readUrl}
\title{Reads urls
Note: taken from a stack overflow thread.
Modify somewhat please.}
\usage{
readUrl(url)

readUrl(url)
}
\value{
Whatever url is read.

Whatever url is read.
}
\description{
Reads urls
Note: taken from a stack overflow thread.
Modify somewhat please.

Reads urls
Note: taken from a stack overflow thread.
Modify somewhat please.
}
