% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noncovid_essentials.R
\name{reference_hcw}
\alias{reference_hcw}
\title{Reference cadres of HCWs}
\usage{
reference_hcw(
  iso3c = NULL,
  params,
  who,
  throughput,
  overrides = list(n_docs = 8000, n_nurses = 5000, n_labs = 300, n_midwives = 500,
    n_dentists = 10, n_physiotherapists = 50, n_trad_comp_med = 4000, n_chws = 245,
    n_pharmacists = 818)
)
}
\arguments{
\item{iso3c}{ISO3C country code.}

\item{params}{From get_parameters}

\item{who}{WHO summary data from the esft package, who.rda}

\item{throughput}{Either diagnostic_capacity or throughput}

\item{overrides}{These are the base values supplied by the ESFT worksheet -
replacements for if there are no reference cadre values}
}
\description{
Reference cadres of HCWs
}
