% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_parameters.R
\name{test_ratio}
\alias{test_ratio}
\title{Get diagnostic ratios}
\usage{
test_ratio(capacity, diagnostic_params)
}
\arguments{
\item{capacity}{From calculate_diagnostic_capacity.}

\item{diagnostic_params}{From get_diagnostic_parameters}
}
\description{
Get diagnostic ratios
}
